inPaste/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-compat==1.0.15
django-cors-headers==4.3.1
django-rest-framework==0.1.0
django4-background-tasks==1.2.9
djangorestframework==3.15.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
